Adds automatically generated colors to Dynmap areas (#11)
These changes add some colors to the Dynmap integration.

The colors can be automatically generated based on the owner's UUID.
That means the same color is generated each time.

Additionally, it is possible to have a colors-file to manually set the colors used per UUID.
This has been requested in #11 and is only recommended for server without many players
as generating the color scales way better and is inexpensive compared to potentially
wasting a lot of MBs in holding the colors file in memory